Restructure jars destination to accommodate many jars
Previously, jar file was copied at the root of the folder
holding binaries.

This was fine with only one jar, but when adding many jar
dependencies, this would make that folder unreadable.

This commit moves the destination of the jars to a specific folder
and enables copying of all jars to it, not only project code.